backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
aiohttp==3.9.1
beautifulsoup4==4.12.2
PyPDF2==3.0.1
google-generativeai==0.3.1
crewai==0.11.0
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
